+++ os_info/src/linux/mod.rs
@@ -33,6 +33,8 @@ mod tests {
             | Type::Debian
             | Type::EndeavourOS
             | Type::Fedora
+            | Type::Garuda
+            | Type::Gentoo
             | Type::Linux
             | Type::Manjaro
             | Type::Mariner