+++ os_info/src/matcher.rs
@@ -33,10 +33,8 @@ impl Matcher {
 fn find_by_key<'a>(string: &'a str, key: &str) -> Option<&'a str> {
     let key = [key, "="].concat();
     for line in string.lines() {
-        if let Some(key_start) = line.find(&key) {
-            return Some(
-                line[key_start + key.len()..].trim_matches(|c: char| c == '"' || c.is_whitespace()),
-            );
+        if line.starts_with(&key) {
+            return Some(line[key.len()..].trim_matches(|c: char| c == '"' || c.is_whitespace()));
         }
     }